Solving rewrite conflicts with php 5.4 trait
