+++ app/code/community/Hackathon/ResponsiveEmail/Model/Email/Template/Ebizmarts.php
@@ -1,0 +1,6 @@
+<?php
+
+class Hackathon_ResponsiveEmail_Model_Email_TemplateEbizmarts extends Ebizmarts_Mandrill_Model_Email_Template
+{
+    use Hackathon_ResponsiveEmail_Model_Email_Template_Trait;
+}
